LUN-6: refactor path to colors
