+++ src/components/ExerciseHeader.tsx
@@ -3,7 +3,7 @@ import Modal from './Modal'
 import { BackHandler, StyleSheet, Text, TouchableOpacity } from 'react-native'
 import { CloseButton } from '../../assets/images'
 import { widthPercentageToDP as wp } from 'react-native-responsive-screen'
-import { COLORS } from '../constants/colors'
+import { COLORS } from '../constants/theme/colors'
 import labels from '../constants/labels.json'
 import { ProgressBar } from 'react-native-paper'
 import { StackNavigationProp } from '@react-navigation/stack'
