+++ src/components/Modal.tsx
@@ -3,7 +3,7 @@ import { View, Text, TouchableOpacity, Modal, StyleSheet } from 'react-native'
 import { CloseIcon } from '../../assets/images'
 import Button from './Button'
 import { BUTTONS_THEME } from '../constants/data'
-import { COLORS } from '../constants/colors'
+import { COLORS } from '../constants/theme/colors'
 import { widthPercentageToDP as wp } from 'react-native-responsive-screen'
 import AsyncStorage from '../services/AsyncStorage'
 import labels from '../constants/labels.json'
