+++ src/navigation/Navigator.tsx
@@ -12,7 +12,7 @@ import InitialSummaryScreen from '../routes/InitialSummaryScreen'
 import ResultsOverviewScreen from '../routes/ResultsOverviewScreen'
 import ResultScreen from '../routes/ResultScreen'
 import { NavigationContainer } from '@react-navigation/native'
-import { COLORS } from '../constants/colors'
+import { COLORS } from '../constants/theme/colors'
 import { widthPercentageToDP as wp } from 'react-native-responsive-screen'
 import WordChoiceExerciseScreen from '../routes/choice-exercises/WordChoiceExerciseScreen'
 import ArticleChoiceExerciseScreen from '../routes/choice-exercises/ArticleChoiceExerciseScreen'
