+++ src/routes/InitialSummaryScreen.tsx
@@ -4,7 +4,7 @@ import Button from '../components/Button'
 import { CheckIcon, ListIcon, RepeatIcon } from '../../assets/images'
 import { BUTTONS_THEME, ExerciseKeys, EXERCISES } from '../constants/data'
 import { RouteProp, useFocusEffect } from '@react-navigation/native'
-import { COLORS } from '../constants/colors'
+import { COLORS } from '../constants/theme/colors'
 import { heightPercentageToDP as hp, widthPercentageToDP as wp } from 'react-native-responsive-screen'
 import { DocumentResultType, RoutesParamsType } from '../navigation/NavigationTypes'
 import { StackNavigationProp } from '@react-navigation/stack'
