+++ src/routes/vocabulary-list/VocabularyListScreen.tsx
@@ -5,7 +5,7 @@ import { DocumentType } from '../../constants/endpoints'
 import Title from '../../components/Title'
 import VocabularyListItem from './components/VocabularyListItem'
 import Loading from '../../components/Loading'
-import { COLORS } from '../../constants/colors'
+import { COLORS } from '../../constants/theme/colors'
 import { widthPercentageToDP as wp } from 'react-native-responsive-screen'
 import { RouteProp } from '@react-navigation/native'
 import { RoutesParamsType } from '../../navigation/NavigationTypes'
