+++ src/routes/vocabulary-list/components/VocabularyListItem.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from 'react'
 import { Image, StyleSheet, Text, View } from 'react-native'
-import { COLORS } from '../../../constants/colors'
+import { COLORS } from '../../../constants/theme/colors'
 import { getArticleColor } from '../../../services/helpers'
 import { widthPercentageToDP as wp } from 'react-native-responsive-screen'
 import AudioPlayer from '../../../components/AudioPlayer'
